# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: LupexWallet API
  version: "1.0.0"
  description: >
    REST API однопользовательского приложения учета личных финансов.
    Технические идентификаторы — на английском, см. docs/architecture/ddd-model.md
    для доменной модели и docs/database/schema.md для схемы БД.
  license:
    name: Proprietary — внутренний проект, не публикуется
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /wallets:
    get:
      tags: [Wallets]
      summary: Список кошельков
      operationId: listWallets
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Страница кошельков, отсортированная по displayOrder
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WalletPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Wallets]
      summary: Создать кошелек (UC-01)
      operationId: createWallet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WalletCreateRequest" }
      responses:
        "201":
          description: Кошелек создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /wallets/{walletId}:
    parameters:
      - $ref: "#/components/parameters/WalletIdParam"
    get:
      tags: [Wallets]
      summary: Получить кошелек
      operationId: getWallet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Wallets]
      summary: Изменить атрибуты кошелька (UC-02, без валюты)
      operationId: updateWallet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WalletUpdateRequest" }
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Wallets]
      summary: Удалить кошелек без истории (UC-04)
      operationId: deleteWallet
      responses:
        "204": { description: Удален }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Кошелек имеет историю операций/слепков (используйте /archive)
            или является основным.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /wallets/{walletId}/archive:
    parameters:
      - $ref: "#/components/parameters/WalletIdParam"
    post:
      tags: [Wallets]
      summary: Архивировать кошелек (UC-03)
      operationId: archiveWallet
      responses:
        "200":
          description: Архивирован
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Нельзя архивировать текущий основной кошелек (решено, Q7).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /wallets/{walletId}/set-primary:
    parameters:
      - $ref: "#/components/parameters/WalletIdParam"
    post:
      tags: [Wallets]
      summary: Назначить кошелек основным (UC-05)
      operationId: setPrimaryWallet
      responses:
        "200":
          description: Назначен основным; include_in_total принудительно включен (Q8)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Нельзя назначить основным архивный кошелек.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /wallets/{walletId}/currency:
    parameters:
      - $ref: "#/components/parameters/WalletIdParam"
    put:
      tags: [Wallets]
      summary: Сменить валюту кошелька — только если нет истории (решено, Q15)
      operationId: changeWalletCurrency
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WalletChangeCurrencyRequest" }
      responses:
        "200":
          description: Валюта изменена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: У кошелька уже есть операции или слепки баланса.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
components:
  parameters:
    WalletIdParam:
      name: walletId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CursorParam:
      name: cursor
      in: query
      description: Непрозрачный курсор для постраничной навигации.
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100, minimum: 1 }
  schemas:
...
    Money:
      type: object
      required: [amount, currencyId]
      properties:
        amount:
          type: string
          format: decimal
          description: Произвольная точность (ADR-0005); передается строкой, чтобы не терять точность в JSON-числах.
          example: "1234.56789012"
        currencyId: { type: string, format: uuid }
        currencyCode: { type: string, readOnly: true, example: "USD" }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    CursorPageMeta:
      type: object
      required: [nextCursor, hasMore]
      properties:
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }

    # ---- Auth ----
...
    Wallet:
      type: object
      required:
        [id, name, walletTypeId, currencyId, initialBalance, accountingStartDate,
         currentBalance, includeInTotal, isPrimary, isArchived, displayOrder,
         createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string, maxLength: 200 }
        walletTypeId: { type: string, format: uuid }
        purposeDescription: { type: [string, "null"] }
        currencyId: { type: string, format: uuid }
        initialBalance: { $ref: "#/components/schemas/Money" }
        accountingStartDate: { type: string, format: date }
        currentBalance: { $ref: "#/components/schemas/Money", readOnly: true }
        includeInTotal: { type: boolean }
        isPrimary: { type: boolean, readOnly: true, description: "Меняется только через POST /wallets/{id}/set-primary" }
        isArchived: { type: boolean, readOnly: true }
        displayOrder: { type: integer }
        color: { type: [string, "null"] }
        icon: { type: [string, "null"] }
        createdAt: { type: string, format: date-time, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
    WalletCreateRequest:
      type: object
      required: [name, walletTypeId, currencyId, initialBalanceAmount, accountingStartDate]
      properties:
        name: { type: string, maxLength: 200 }
        walletTypeId: { type: string, format: uuid }
        purposeDescription: { type: [string, "null"] }
        currencyId: { type: string, format: uuid }
        initialBalanceAmount: { type: string, format: decimal }
        accountingStartDate: { type: string, format: date }
        includeInTotal: { type: boolean, default: true }
        displayOrder: { type: integer, default: 0 }
        color: { type: [string, "null"] }
        icon: { type: [string, "null"] }
    WalletUpdateRequest:
      type: object
      required: [name, walletTypeId, includeInTotal, displayOrder]
      properties:
        name: { type: string, maxLength: 200 }
        walletTypeId: { type: string, format: uuid }
        purposeDescription: { type: [string, "null"] }
        includeInTotal:
          type: boolean
          description: Игнорируется (остается true), если кошелек основной (Q8).
        displayOrder: { type: integer }
        color: { type: [string, "null"] }
        icon: { type: [string, "null"] }
    WalletChangeCurrencyRequest:
      type: object
      required: [currencyId]
      properties:
        currencyId: { type: string, format: uuid }
    WalletPage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Wallet" } }
        pagination: { $ref: "#/components/schemas/CursorPageMeta" }

    # ---- Reference data ----
...
  responses:
    BadRequest:
      description: Некорректные параметры запроса
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Отсутствует или неверен токен доступа
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Запись не найдена
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: Конфликт с текущим состоянием (нарушение инварианта)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT